Adding arping from orchestrator server
This is because I found that the arp tables are taking longer to refresh. So, in a failover and fail back, I found that the vip was not being shutdown. This is weird and I was not able to explain why it is not able to knowledge that the vip has changed, but the workaround is simple and worked.